Prepare FFI for the merged register-file mode
The debug reporter's extern for validation_failure_scope now matches the
generated register's uint8_t type instead of over-reading as uint64_t,
and EXECUTE_STACK pastes its handler name directly so an argument
spelled like a register (pc) cannot be macro-expanded before pasting
once the register-file compatibility macros are active.

+++ zkvm/zisk/platform.c
@@ -39,7 +39,7 @@ void htif_puts(const char *s)
 
 #ifdef EVMSAIL_DEBUG
 extern _Bool validation_failure_present;
-extern uint64_t validation_failure_scope;
+extern uint8_t validation_failure_scope;
 extern int validation_failure_reason;
 
 static char hex_digit(unsigned value)
